removed the archive
### How to install
This launcher will require some dependencies to run, so you cant just run the .app file right of the bat. The file comes bundled with an installer binary. Simply double click it to start the installation.
*If terminal doesn't open when you double-click it, open Terminal and type "chmod +x " and drag the installer file into terminal. Press enter and then try to double-click the installer.*

If the installer doesn't work, make sure you have **git** installed.

I'm not sure how this will run on other computers.

this is what was removed.